fix(appsec): report all tags on the service entry span instead of the local root span (#14210)
## Motivation

In specific cases where we have an inferred span (belonging to an
inferred service) as the parent of the framework instrumented web
request (the 'web' span), Appsec is enabled on the inferred service
instead of the current service.

This is because we always use set tags and metrics on the root span
using the `span._local_root` helper.

This leads to an incorrect reporting of the appsec enablement status,
and security signals showing up on spans of a different service.

Note:
- The only type of inferred span in my understanding are API Gateways.
They can be created either through the AWS Lambda python layer and are
of type 'http' or are inferred by dd-trace-py and are of type 'web'.
- Appsec through the tracer in lambda is not generally available yet,
and the extension correctly handles this case. So no problems for now.
But is necessary for in tracer enablement of appsec.

## Changes

- Add a helper on spans to retrieve the top level span of a given
service

- Modify the appsec logic to always report on the service entry span and
add a failsafe to query the entry span of the current span or current
root span to get a handle on a span in any situation.

- Modify the tests to look for the entry span instead of the root span
 
 Notes:
- for IAST, we still need to check the root span in some cases because
of the `_DD_IAST_USE_ROOT_SPAN` flag. This setting also needs additional
manual handling in `report_stack` as reporting to root span is not the
default anymore.
- The contrib_appsec tests contain shell injections (that we ensure are
detected by appsec). I included that path to the codeql bypass file.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)