fix(er): start capturing from the leaf frame
The introduction of the maximum number of frames to capture in Exception
Replay has caused the capture to be truncated at the outermost exception
end, close to the exception handler frame. This follows the natural
ordering of traceback frames in traceback objects, which go in the
opposite direction of stack traces. We reverse the order of the frames
in an exception chain to ensure that we start capturing from the leaf
frame of the innermost exception, thus starting from the end where the
exception was actually thrown. We also ensure to keep the numbering
consistent with the current implementation to allow for the correct
matching on the UI side.