feat(llmobs): langchain integration prompt template tracking (#14107)
Adding prompt template tagging for langchain integration.

Some thoughts and clarifications:
The overall flow here is to capture the template data on template
invocation (BasePromptTemplate.invoke()), and attach it to the resulting
PromptValue instance.
Then, on LLM.invoke(), carry that template data forward onto the the the
LLM call instance.
Then, on LLM.generate(), where the span is finally created, attach it to
the span.

One complication around langchain is that the final llm span is often
not created with the langchain integration, but with a provider
integration (openai, anthropic, etc). Thus, when we do the actual span
construction, we may need to look not at the span tags, but at the
parent span tags to find the prompt object.

Also, since we want to construct a consistent promptId across template
instances, we try here to use reflection to find a variable and module
name. For some situations, it may be that we fail, such as when the
prompt template is only ever an rvalue.

Not included here is the ChatTemplate prompt templates. This is because
the updated Prompt type that allows for non-string templates is not yet
merged. I will revisit these once that type is in.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>
Co-authored-by: Sam Brenner <106700075+sabrenner@users.noreply.github.com>