perf(profiler): don't start an extra thread for memory profiling (#14288)
The MemoryCollector inherits from the PeriodicService class, which in turn
starts a thread to periodically call a given Python callback. The callback is
the periodic method, which for MemoryCollector is just the no-op method
inherited from PeriodicService.

I wasn't even aware of this behavior just looking at MemoryCollector, which
caused problems when I deleted the `_interval` member variable in a PR.
After that, the periodic thread ran in a tight loop. This is how I learned that the
thread was even there... Really we don't need the periodic thread for the
memory profiler. The main profiling thread is already responsible for setup,
teardown, and collecting periodic snapshots of the memory profile.