chore(profiling): deflake test_memory_collector_thread_lifecycle (#14368)
The test is running with quite a high sampling interval, but is doing
allocations we mostly don't sample right now (list objects where the
backing memory allocation isn't profiled). As a result, there is a
decent chance that we don't sample allocations on the "worker" threads
in the test since they do ~560 bytes of sampleable allocations and the
interval is 512. Decrease the sample interval to 8 bytes so that we're
basically recording every allocation.