perf: use lru_cache for caching
All the Python versions that we support offer lru_cache from the
standard library. We can use this instead of our own implementation
because it is generally implemented in native and offers better
performance, even though it is not specialised to the single argument
case of our internal implementation. A simple test shows a 6x overhead
reduction when opting for the stdlib solution.