fix(profiling): avoid use-after-free when profiling realloc
Prior to Python 3.12, and particularly in Python 3.11, collecting tracebacks
while intercepting allocations is prone to crashes. We currently use the C
Python API to collect tracebacks, which can do allocations. These allocations
can in turn trigger garbage collection, allowing other code to run. In the past
we've seen this lead to GIL release and cause corruption in the memory
profiler.

This can also lead to use-after-free crashes. For example, calling realloc to
grow a data structure, we can trigger garbage collection which visits the data
structure. The underlying reallocaiton will have already happened (we want to
track the new address) but the data structure will still point to old memory
since our wrapper hasn't returned.

See, for example, the elementtree module.
https://github.com/python/cpython/blob/v3.11.13/Modules/_elementtree.c#L465
reallocates memory, and the reallocated field is visited here during GC:
https://github.com/python/cpython/blob/v3.11.13/Modules/_elementtree.c#L634

Python 3.12 doesn't trigger GC during allocation. Instead, a flag is set for GC
to run later, at a safe point in the interpreter. But for earlier versions, we
disable GC temporarily. This will allow a small, temporary increase in memory
usage during sampling. But it is overall cheap (mostly just toggling a boolean)
and the alternative is hard-to-diagnose crashes.