chore(debugger): only cache agent info for 60s (#14568)
Cache agent information for up to 60 seconds to handle environments
where the agent version may vary.

In some setups, agents run behind a load balancer. A request might first
hit a newer agent with the required endpoints, but a subsequent request
could be routed to an older version. Invalidating the cache will help
protect resolve broken scenarios.

There's an open question around locking out of newer endpoints when we
receive a 4xx but that seems like overkill for now.