fix(di): get pending probe as a copy
When retrieving the list of pending probes, we make the registry return
a copy of the collection to avoid mutation during iteration. In the
worst case, Python does not warn that the list has been mutated while
iterating over it, causing the iteration loop to end prematurely without
any sings of errors.