ci: pre-compute library version for build steps (#14595)
This PR adds an additional stage to the GitHub Action build workflow, as
well as our GitLab job to pre-compute the setuptools_scm version we are
going to build.

Right now when we build in GitHub Actions we get a different
setuptools_scm version than in GitLab, this is because GitHub Actions
operates on a merge commit, and GitLab on the actual commit. This makes
it so if we ever wanted to migrate any building to GitLab it has to be
an all or nothing approach since versions will differ.

This Change makes it so we can compute and store the intended version
once at the start of the workflow such that we re-use the same version
throughout the whole pipeline, and be able to download the version into
GitLab to re-use there for consistency.

We have an additional change/benefit. When computing the version, if we
are on the main branch, a `[0-9].x` branch, or a `[0-9].[0-9]+` branch,
then we will compute using `setuptools-scm --strip-dev` which removes
the dev suffix from the version. This means the artifacts build on those
branches will use the next actual version we intend to release, e.g.
`3.15.0` on main right now, `3.14.2` on `3.14`, etc. This is a nice
bonus for us.


We added a new job to GitLab to download the version there for use in
the pipeline, but it isn't currently being used by anything. I wanted to
keep the scope of this PR smaller by only adding it as a job to the
pipeline / making the version available to prove we can do it and set up
the base.


## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)