fix(er): proactively remove exception fields (#14649)
## Description

Exception Replay introduces an unexpected interaction with Celery when
tasks raise custom exceptions with mandatory arguments. It seems that
holding on to exception fields, such as __cause__, for longer than
necessary causes Celery to produce bad pickled responses that are not
handled by the framework. We therefore proactively clear up our tracking
of the exception fields that we want to capture to prevent Celery tasks
from crashing the application.