feat(debugger): add agent state check to uploader (#14653)
## Description

We add an agent state check to the uploader to allow it to fall back to
an appropriate snapshot collection endpoint that is guaranteed to have
redaction support. If such guarantees cannot be made we log a warning
and disable the uploads of snapshots to prevent sensitive data from
being captured.

<img width="784" height="373" alt="image"
src="https://github.com/user-attachments/assets/6c51aba1-185d-42d4-ac33-96f2cf0b128d"
/>

---------

Co-authored-by: Tyler Finethy <tylfin@gmail.com>