chore: avoid building profiling extension under 3.14 (#14677)
This change adds Python 3.14 to the list of runtime versions available
in the shell for some gitlab and github workflows. It also skips
building of the profiling extension under 3.14, since the profiling
product currently doesn't work with this version.

Behavior under 3.14 isn't validated in tests - for now it's enough to
know that this change doesn't break any existing tests.
This was pulled out of
https://github.com/DataDog/dd-trace-py/pull/14264, which will include
automated testing of this change under 3.14.