test: remove checks for no error logs from many tests (#14670)
## Description

This change removes checks for the lack of error logs from many tests
that run ddtrace in a subprocess. There are some tests in the suite that
directly validate the contents of stderr. The ones changed here are
primarily checking other behaviors and break when unrelated code writes
to stderr.