chore(telemetry): telemetry log messages must be constant (#14642)
Last PR in a list of 3 PRs to achieved requirements listed
[here](https://docs.google.com/document/d/16SFeOp7yxzT6dM1O0TKtWkgDsiq6SeV8P_MAYboCdOI/edit?tab=t.gho5ikpq9hed).

See: https://github.com/DataDog/dd-trace-py/pull/14639.
See: #14641 

# Constant log message enforcement

Every log.error() messages are by default reported to telemetry. Message
reported to telemetry must be constant.

A check was added in CI (you can test it by running `hatch run
lint:error-log-check`): you will not be able to merge a PR that sends a
non constant message to telemetry.
You are two ways to still doing that:
- `log.error("my message", extra={"send_to_telemetry"=False})`: this
will only log the message and not sending it to telemetry.
- add an exception in `check_constant_log_message.py`. The reason should
be added in a comment and is only valid if you are 100% that the message
is constant or the log.error will not send telemetry. For instance in a
file of the debugger, log.error will eventually call log.debug.

Note that `log.error("my message with: %s", "something")` is valid but
only "my messag with %s", will be sent to telemetry.

## Checklist
- [X] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)