fix(aap): fix flask patch for none methods (#14691)
## Description

Some problems were found on dogweb with flask.
This fix ensures the `methods` value is properly iterable.

Also: use the default "GET" method as specified in the Flask
documentation in case no methods have been defined.

## Testing

Add a specific endpoint with methods=None to the flask application for
appsec_threats test as a regression test, triggering the bug in previous
versions of the tracer.

APPSEC-59242