fix(llmobs): ensure DD_APM_TRACING_ENABLED disables APM (#14617)
The existing DD_APM_TRACING_ENABLED functionality is specially
implemented for asm and did not work for llmobs. Spans are still sent
since some spans are required for asm.

This change introduces a trace filter to drop all apm spans when llmobs
is enabled.

Consequently, asm + llmobs with apm disabled will not be possible with
this change. I'm choosing to avoid complexity in this implementation
until we see a concrete ask for this functionality.

Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>