ci: actually validate the ast-grep rules against the repo (#14711)
## Description

We added ast-grep and a custom rule for checking `ddtrace.trace.Pin`
usage and warning that it is deprecated.

We updated CI to run our ast-grep pattern test suite (tests that our
rules work), but we never actually ran the "scan" which validates the
repo for the rules.

## Testing

There is 1 known violation in the codebase, I opened the draft PR first
with adding the `hatch run lint:sg` to ensure it caught the issue first
before fixing the violation.

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->