chore(llmobs): add multi message prompt templates support for langchain (#14671)
## Description
https://python.langchain.com/api_reference/core/prompts.html#
Recently added support for auto-instrumentation of prompt templates in
langchain, but the way it was done only worked when the templates were
simple strings, due to the constraints of our prompt template type/back
compatibility.
following [this pr](https://github.com/DataDog/dd-trace-py/pull/12551),
those issues are dealt with. here, we fix that original half-effort and
enable instrumentation that can properly support multi-message
templates.
this is a shadow release, while the FE and event handling finishes, thus
no changelog

known issue with the reflection prompt id scheme: in certain lcel
situations, as well as certain batched situations, seems like the
reflection isn't working. wondering if it has to do with some threading
langchain may be doing?
if so, one approach could be to wrap __init__, and grab what we can
there. but the variable isn't assigned at that point, so we can probably
only get module name (and the enclosing function name maybe?).
a hybrid approach might be to take the current at-invocation time method
when it works, and fall back on the module_name.function_name? open to
ideas.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

---------

Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>