fix(llmobs): [MLOB-3941] parse model and provider from inference profiles (#14713)
## Description

See the corresponding JIRA ticket for details on the customer report of
this bug. Basically, we aren't guaranteed to be able to parse the model
/ provider from a Bedrock model ID. This PR fixes one of the issues we
had which was not handling inference profile model ID formats properly.
Additionally, I introduced a method to do a last ditch attempt at
matching one of our known provider strings to the model ID. If this
fails, we default to custom which is the current behavior.

As an aside, I also looked into whether it was possible to extract the
model / provider from the response body instead; however, it doesn't
seem like the norm for the model to be included in the response (e.g.
Amazon Titan, A121 Jurassic models both did not include the model in
their responses).

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->