feat(ray): add ray core integration (#14511)
This PR adds an integration for [Ray
Core](https://docs.ray.io/en/latest/ray-core/walkthrough.html).
> Ray Core is a powerful distributed computing framework that provides a
small set of essential primitives (tasks, actors, and objects) for
building and scaling distributed applications.

##  Ray Overview
This is a simple ray example:
```python
import ray

@ray.remote
def square(x):
    result = x * x + 1
    return result

futures = [square.remote(i) for i in range(3)]

print(ray.get(futures))
```

When executing this job/submitting this job here is what happens:
<img width="1522" height="1232" alt="image"
src="https://github.com/user-attachments/assets/dc74a58d-89cc-4da4-a849-8ce69d0d4942"
/>

Tasks are stateless called `RemoteFunction` in the ray code. There is
something a similar concept called `Actor` which are stateful. See the
link above for more details on Ray Core.

## What this PR wants to achieve

From the Ray code we are able to know: 
- When a job is submitted
- When a remote function is created and called remotely (we are
submitting a task)
- When an Actor is created and when an actor method is called remotely
(we are submitting an actor method)
- When a job is finished

However, we also need:
- Info on when a task is executed
- Info on when an actor method is executed 
- A root span which lasts from submission to end of execution

###  Constraints

- A job will be executed across several processes. These processes are
handled by C++ code and we have no control on communications from
python. There are no network calls as well between them
- We have very little controls about processes lifetime, for example
workers/actors can be killed/garbage collected from C++ code
- Ray jobs/tasks can be very long (several hours, even several days).

## What this PR does

This PR adds tracing for:
- Task submission and execution
- Actor method submission and execution
- Job lifetime (job submit/job start/job end)

### Tracing task/actor method execution

Task/Actor method are user code that we cannot know in advance. To trace
task execution we wrap the function with tracing code when a task is
submitted. At that time a remote function is created and we can know
what function will be executed.
To trace actor method execution, we wrap all the function of the Actor
class when it is created.

### Communication across processes

To propagate context across processes we are using two mechanisms:
- Injecting context within env variables. We inject context from some
spans into the environment variables at the job submission level and
when the job starts. The env variables are propagated by ray which
allows us to retrieve them in other processes.
- When a task is submitted, we can modify the signature of the function
that will be executed and inject the tracing context in the kwargs of
the function so we can have parent/child relationship between task
submission and task execution which are done in different processes. We
do the same for actor method.

### Process lifetime 

The only way to ensure that all the spans will be sent before a
worker/actor process is killed is to activate sync mode like for
serverless environment.
I tried to add a `time.sleep` before an actor is garbage collected or
before a worker is disconnected. It worked but it was not robust and
added minimum 1.5s (for one actor, one worker) to your job execution.

### Long running job

If a job lasts 1 day, the user wants to know it is running before it is
finished. To solve this issue, this PR adds long running span support.
When we create a span that might be long, we start it as long running
span:
- If it lasts less then 10 seconds, there is no difference with a normal
span.
- If it lasts more then 10 seconds, the span will be sent. Then every
120s, we resubmit the span until its completion.

This case is supported by APM, see [Running
spans](https://datadoghq.atlassian.net/wiki/spaces/ATI/pages/4433936963/Running+spans).

The tracer don't deal well with multiple finishes (it breaks the
context). Therefore we finish the real span only once, the spans emitted
before the real finish are copies.
This causes children of long running spans not being sent until the
parent is finished. To solve that, when we resubmit a long running span,
we also send all the children spans which are finished.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Yakov Shapiro <yakov.shapiro@datadoghq.com>
Co-authored-by: savitagm <savita.manghnani@datadoghq.com>