fix(ci_visibility): include imported modules in test code coverage (#14719)
CI Visibility: This fix resolves an issue where code imported at module
level but not executed during a test would
not be considered by Test Impact Analysis as impacting the test. For
example, a test using a constant imported from
some other module would not count the constant definition among its
impacting lines, because the constant definition
is not executed during the test, but rather when the module was
imported. With this change, code executed at import
time is also included among the impacted lines of a test.

Added unit test for the imported constant case. Modified existing
coverage tests, as now they include more covered lines.

Also did some manual testing with a simple repo containing a test
importing a constant from a different module and checking that a change
in the constant triggers the test to be run.

We now include more lines among the impacted lines of a test, which
means we will skip fewer tests. All lines executed at the top level of
the imported modules are included, which is a bit coarse grained and
means we might run tests that could theoretically be skipped, but this
is safer than the opposite (skipping tests that should be run). Also it
does not really matter because the backend does not care about the
specific lines covered, only about the filenames covered by a test.

None.

(cherry picked from commit 9394842476cfa230a5914d3d27016d6f4711176d)