chore(ci_visibility): add file-cache layer for api responses (#14717)
## Description
CI Visibility: The CI Visibility service makes several requests to the
backend API (depending on the features enabled for the repo). Since the
responses don't usually change, except some specific settings flow that
can require two requests, we can reuse the API responses if we store
them. This becomes especially interesting when xdist, or any other
multi-process environment is used, as we manage to reduce the API calls
from N (one per process) to 1, saving time and bandwith, hence having
the cache as files instead of memory.

The cache is deleted with an `atexit` hook, so once the process finishes
they are removed.

This PR introduces the caching mechanism, but it is disabled by default,
and can be enabled with the env var
`_DD_CIVISIBILITY_RESPONSE_CACHE_ENABLED=1`

## Testing

Unit tests

## Risks

None

---------

Co-authored-by: Vítor De Araújo <vitor.dearaujo@datadoghq.com>