chore: initialise HTTP client when cloning modules
A recent attempt to reduce start-up time by lazy-loading the HTTP client
modules resulted in a potential gevent support regression whereby the
lazily created HTTP client could end up making requests via the
gevent-monkey-patched socket module. This introduces the possibility of
spurious deadlock during the processing of requests. We prevent this by
force-importing the HTTP clients on bootstrap to ensure that we grab
references to an original copy of the socket module. This prevents our
HTTP clients from calling into gevent when uploading payloads to the
agent, preventing the potential deadlocks.