chore(serverless): test to ensure that all imports in datadog-lambda-python still valid (#14746)
## Description

<!-- Provide an overview of the change and motivation for the change -->

Customer reported that a class name changed in the most recent version
of ddtrace which was then causing errors when attempting to import
`datadog_lambda`. See
https://github.com/DataDog/datadog-lambda-python/pull/661 and
https://github.com/DataDog/dd-trace-py/pull/14653.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

It's just some tests, so risk is low.

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

Added additional tests on the `datadog_lambda` side as well, including
running our unit tests twice a day. See
https://github.com/DataDog/datadog-lambda-python/pull/662.