perf(tracer): avoid unncessary context creation (#14773)
## Description

Currently for a new root span we are creating 2 context objects. We
create an empty one in `Tracer._start_span`, and then it gets copied
when we create the root span.

This is unnecessary, we can avoid this extra object allocation/GC.

Benchmarks showing about a 12-15% reduction in root span creation
overhead (only for starting the span, the span-finish scenario is 4-5%
faster).

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->