fix(ci_visibility): do not use is_user_code (#14794)
## Description

Currently, in Test Impact Analysis code coverage, we use
`is_user_code()` to decide whether to instrument a module. This function
was conceived for use in non-CI contexts where user code may be
installed in site-packages like external dependencies. This is not
necessary in Test Optimization, where we only want to cover code inside
the current repo (TIA only works with filenames that are relative to the
repo root, and we already set the repo root as the include path of
`ModuleCodeCollector`). Removing this check also resolves a bug where
test packages are incorrectly detected as third-party if a third-party
package contains files under `test/*` or `tests/*` that get installed at
site-packages level (`sumy` and `sendgrid` do this, for instance).

## Testing

Unit tests + manual testing in dogweb.

## Risks

Potentially covering more code than expected. Should not be an issue
because we already exclude site-packages and similar
[here](https://github.com/DataDog/dd-trace-py/blob/vitor-de-araujo%2FSDTEST-2747%2Fcoverage-third-party/ddtrace/internal/coverage/code.py#L53).

## Additional Notes

None.